+++ warp/builtins.py
@@ -1212,20 +1212,69 @@ add_builtin(
 )
 
 
-# not making these functions available outside kernels (export=False) as they
-# return data via references, which we don't currently support:
+def svd3_value_func(arg_types: Mapping[str, type], arg_values: Mapping[str, Any]):
+    if arg_types is None:
+        return (
+            matrix(shape=(3, 3), dtype=Float),
+            vector(length=3, dtype=Float),
+            matrix(shape=(3, 3), dtype=Float),
+        )
+
+    dtype = arg_types["A"]._wp_scalar_type_
+    return (
+        matrix(shape=(3, 3), dtype=dtype),
+        vector(length=3, dtype=dtype),
+        matrix(shape=(3, 3), dtype=dtype),
+    )
+
+
+add_builtin(
+    "svd3",
+    input_types={"A": matrix(shape=(3, 3), dtype=Float)},
+    value_func=svd3_value_func,
+    group="Vector Math",
+    doc="""Compute the SVD of a 3x3 matrix ``A``. The singular values are returned in ``sigma``,
+    while the left and right basis vectors are returned in ``U`` and ``V``.""",
+)
+
 add_builtin(
     "svd3",
     input_types={
         "A": matrix(shape=(3, 3), dtype=Float),
         "U": matrix(shape=(3, 3), dtype=Float),
         "sigma": vector(length=3, dtype=Float),
-        "V": matrix(shape=(3, 3), dtype=Scalar),
+        "V": matrix(shape=(3, 3), dtype=Float),
     },
     value_type=None,
     group="Vector Math",
     export=False,
     doc="""Compute the SVD of a 3x3 matrix ``A``. The singular values are returned in ``sigma``,
+    while the left and right basis vectors are returned in ``U`` and ``V``.""",
+)
+
+
+def svd2_value_func(arg_types: Mapping[str, type], arg_values: Mapping[str, Any]):
+    if arg_types is None:
+        return (
+            matrix(shape=(2, 2), dtype=Float),
+            vector(length=2, dtype=Float),
+            matrix(shape=(2, 2), dtype=Float),
+        )
+
+    dtype = arg_types["A"]._wp_scalar_type_
+    return (
+        matrix(shape=(2, 2), dtype=dtype),
+        vector(length=2, dtype=dtype),
+        matrix(shape=(2, 2), dtype=dtype),
+    )
+
+
+add_builtin(
+    "svd2",
+    input_types={"A": matrix(shape=(2, 2), dtype=Float)},
+    value_func=svd2_value_func,
+    group="Vector Math",
+    doc="""Compute the SVD of a 2x2 matrix ``A``. The singular values are returned in ``sigma``,
     while the left and right basis vectors are returned in ``U`` and ``V``.""",
 )
 
@@ -1235,13 +1284,37 @@ add_builtin(
         "A": matrix(shape=(2, 2), dtype=Float),
         "U": matrix(shape=(2, 2), dtype=Float),
         "sigma": vector(length=2, dtype=Float),
-        "V": matrix(shape=(2, 2), dtype=Scalar),
+        "V": matrix(shape=(2, 2), dtype=Float),
     },
     value_type=None,
     group="Vector Math",
     export=False,
     doc="""Compute the SVD of a 2x2 matrix ``A``. The singular values are returned in ``sigma``,
     while the left and right basis vectors are returned in ``U`` and ``V``.""",
+)
+
+
+def qr3_value_func(arg_types: Mapping[str, type], arg_values: Mapping[str, Any]):
+    if arg_types is None:
+        return (
+            matrix(shape=(3, 3), dtype=Float),
+            matrix(shape=(3, 3), dtype=Float),
+        )
+
+    dtype = arg_types["A"]._wp_scalar_type_
+    return (
+        matrix(shape=(3, 3), dtype=dtype),
+        matrix(shape=(3, 3), dtype=dtype),
+    )
+
+
+add_builtin(
+    "qr3",
+    input_types={"A": matrix(shape=(3, 3), dtype=Float)},
+    value_func=qr3_value_func,
+    group="Vector Math",
+    doc="""Compute the QR decomposition of a 3x3 matrix ``A``. The orthogonal matrix is returned in ``Q``,
+    while the upper triangular matrix is returned in ``R``.""",
 )
 
 add_builtin(
@@ -1256,6 +1329,27 @@ add_builtin(
     export=False,
     doc="""Compute the QR decomposition of a 3x3 matrix ``A``. The orthogonal matrix is returned in ``Q``,
     while the upper triangular matrix is returned in ``R``.""",
+)
+
+
+def eig3_value_func(arg_types: Mapping[str, type], arg_values: Mapping[str, Any]):
+    if arg_types is None:
+        return (matrix(shape=(3, 3), dtype=Float), vector(length=3, dtype=Float))
+
+    dtype = arg_types["A"]._wp_scalar_type_
+    return (
+        matrix(shape=(3, 3), dtype=dtype),
+        vector(length=3, dtype=dtype),
+    )
+
+
+add_builtin(
+    "eig3",
+    input_types={"A": matrix(shape=(3, 3), dtype=Float)},
+    value_func=eig3_value_func,
+    group="Vector Math",
+    doc="""Compute the eigendecomposition of a 3x3 matrix ``A``. The eigenvectors are returned as the columns of ``Q``,
+    while the corresponding eigenvalues are returned in ``d``.""",
 )
 
 add_builtin(
@@ -1423,13 +1517,34 @@ add_builtin(
     group="Quaternion Math",
     doc="Construct a quaternion representing a rotation of angle radians around the given axis.",
 )
+
+
+def quat_to_axis_angle_value_func(arg_types: Mapping[str, type], arg_values: Mapping[str, Any]):
+    if arg_types is None:
+        return (vector(length=3, dtype=Float), Float)
+
+    dtype = arg_types["quat"]._wp_scalar_type_
+    return (vector(length=3, dtype=dtype), dtype)
+
+
+add_builtin(
+    "quat_to_axis_angle",
+    input_types={"quat": quaternion(dtype=Float)},
+    value_func=quat_to_axis_angle_value_func,
+    group="Quaternion Math",
+    doc="Extract the rotation axis and angle radians a quaternion represents.",
+)
+
 add_builtin(
     "quat_to_axis_angle",
     input_types={"quat": quaternion(dtype=Float), "axis": vector(length=3, dtype=Float), "angle": Float},
     value_type=None,
     group="Quaternion Math",
     doc="Extract the rotation axis and angle radians a quaternion represents.",
+    export=False,
 )
+
+
 add_builtin(
     "quat_from_matrix",
     input_types={"mat": matrix(shape=(3, 3), dtype=Float)},
@@ -1554,7 +1669,6 @@ add_builtin(
 
 
 def transform_identity_value_func(arg_types: Mapping[str, type], arg_values: Mapping[str, Any]):
-    # if arg_types is None then we are in 'export' mode
     if arg_types is None:
         # return transformation(dtype=Float)
         return transformf
@@ -4146,6 +4260,7 @@ add_builtin(
     input_types={"id": uint64, "i": int, "j": int, "k": int, "value": float},
     group="Volumes",
     doc="""Store ``value`` at the voxel with coordinates ``i``, ``j``, ``k``.""",
+    export=False,
 )
 
 add_builtin(
@@ -4173,6 +4288,7 @@ add_builtin(
     input_types={"id": uint64, "i": int, "j": int, "k": int, "value": vec3},
     group="Volumes",
     doc="""Store ``value`` at the voxel with coordinates ``i``, ``j``, ``k``.""",
+    export=False,
 )
 
 add_builtin(
@@ -4198,6 +4314,7 @@ add_builtin(
     input_types={"id": uint64, "i": int, "j": int, "k": int, "value": int},
     group="Volumes",
     doc="""Store ``value`` at the voxel with coordinates ``i``, ``j``, ``k``.""",
+    export=False,
 )
 
 
@@ -5341,6 +5458,7 @@ add_builtin(
     input_types={"a": vector(length=Any, dtype=Scalar), "i": int, "value": Scalar},
     value_type=None,
     hidden=True,
+    export=False,
     group="Utility",
 )
 
@@ -5350,6 +5468,7 @@ add_builtin(
     input_types={"a": quaternion(dtype=Scalar), "i": int, "value": Scalar},
     value_type=None,
     hidden=True,
+    export=False,
     group="Utility",
 )
 
@@ -5383,6 +5502,7 @@ add_builtin(
     input_types={"a": vector(length=Any, dtype=Scalar), "i": int, "value": Scalar},
     value_type=None,
     hidden=True,
+    export=False,
     group="Utility",
 )
 
@@ -5392,6 +5512,7 @@ add_builtin(
     input_types={"a": quaternion(dtype=Scalar), "i": int, "value": Scalar},
     value_type=None,
     hidden=True,
+    export=False,
     group="Utility",
 )
 
@@ -5401,6 +5522,7 @@ add_builtin(
     input_types={"a": vector(length=Any, dtype=Scalar), "i": int, "value": Scalar},
     value_type=None,
     hidden=True,
+    export=False,
     group="Utility",
 )
 
@@ -5410,6 +5532,7 @@ add_builtin(
     input_types={"a": quaternion(dtype=Scalar), "i": int, "value": Scalar},
     value_type=None,
     hidden=True,
+    export=False,
     group="Utility",
 )
 
@@ -5464,6 +5587,7 @@ add_builtin(
     input_types={"a": matrix(shape=(Any, Any), dtype=Scalar), "i": int, "j": int, "value": Scalar},
     value_type=None,
     hidden=True,
+    export=False,
     group="Utility",
 )
 
@@ -5475,6 +5599,7 @@ add_builtin(
     constraint=matrix_vector_sametype,
     value_type=None,
     hidden=True,
+    export=False,
     group="Utility",
 )
 
@@ -5511,6 +5636,7 @@ add_builtin(
     input_types={"a": matrix(shape=(Any, Any), dtype=Scalar), "i": int, "j": int, "value": Scalar},
     value_type=None,
     hidden=True,
+    export=False,
     group="Utility",
 )
 
@@ -5522,6 +5648,7 @@ add_builtin(
     constraint=matrix_vector_sametype,
     value_type=None,
     hidden=True,
+    export=False,
     group="Utility",
 )
 
@@ -5532,6 +5659,7 @@ add_builtin(
     input_types={"a": matrix(shape=(Any, Any), dtype=Scalar), "i": int, "j": int, "value": Scalar},
     value_type=None,
     hidden=True,
+    export=False,
     group="Utility",
 )
 
@@ -5542,6 +5670,7 @@ add_builtin(
     input_types={"a": matrix(shape=(Any, Any), dtype=Scalar), "i": int, "value": vector(length=Any, dtype=Scalar)},
     value_type=None,
     hidden=True,
+    export=False,
     group="Utility",
 )
 
@@ -5566,6 +5695,7 @@ for t in scalar_types + vector_types + (bool,):
         doc="Prints an error to stdout if ``a`` and ``b`` are not equal",
         group="Utility",
         hidden=True,
+        export=False,
     )
 
 
@@ -5593,6 +5723,7 @@ add_builtin(
     doc="Prints an error to stdout if ``a`` and ``b`` are equal",
     group="Utility",
     hidden=True,
+    export=False,
 )
 
 add_builtin(
@@ -5612,6 +5743,7 @@ add_builtin(
     doc="Prints an error to stdout if ``a`` and ``b`` are equal",
     group="Utility",
     hidden=True,
+    export=False,
 )
 
 add_builtin(
@@ -5682,11 +5814,23 @@ add_builtin(
     group="Utility",
 )
 
+
 # fuzzy compare for float values
+def expect_near_constraint(arg_types: Mapping[str, type]):
+    if not types_equal(arg_types["a"], arg_types["b"]):
+        return False
+
+    if hasattr(arg_types["a"], "_wp_scalar_type_"):
+        return types_equal(arg_types["a"]._wp_scalar_type_, arg_types["tolerance"])
+
+    return types_equal(arg_types["a"], arg_types["tolerance"])
+
+
 add_builtin(
     "expect_near",
     input_types={"a": Float, "b": Float, "tolerance": Float},
     defaults={"tolerance": 1.0e-6},
+    constraint=expect_near_constraint,
     value_type=None,
     doc="Prints an error to stdout if ``a`` and ``b`` are not closer than tolerance in magnitude",
     group="Utility",
@@ -5695,6 +5839,7 @@ add_builtin(
     "expect_near",
     input_types={"a": vector(length=Any, dtype=Float), "b": vector(length=Any, dtype=Float), "tolerance": Float},
     defaults={"tolerance": 1.0e-6},
+    constraint=expect_near_constraint,
     value_type=None,
     doc="Prints an error to stdout if any element of ``a`` and ``b`` are not closer than tolerance in magnitude",
     group="Utility",
@@ -5703,6 +5848,7 @@ add_builtin(
     "expect_near",
     input_types={"a": quaternion(dtype=Float), "b": quaternion(dtype=Float), "tolerance": Float},
     defaults={"tolerance": 1.0e-6},
+    constraint=expect_near_constraint,
     value_type=None,
     doc="Prints an error to stdout if any element of ``a`` and ``b`` are not closer than tolerance in magnitude",
     group="Utility",
@@ -5715,6 +5861,7 @@ add_builtin(
         "tolerance": Float,
     },
     defaults={"tolerance": 1.0e-6},
+    constraint=expect_near_constraint,
     value_type=None,
     doc="Prints an error to stdout if any element of ``a`` and ``b`` are not closer than tolerance in magnitude",
     group="Utility",